更新桌面配置和readme

+++ config_ubuntu_desktop.sh
@@ -15,16 +15,26 @@ sudo update-grub
 sudo sed -i 's/#DefaultTimeoutStopSec=90s/DefaultTimeoutStopSec=5s/' /etc/systemd/system.conf
 systemctl daemon-reload
 if [[ "$XDG_CURRENT_DESKTOP" == "ubuntu:GNOME" ]]; then
-    sudo apt-fast install -y gnome-tweaks gnome-shell-extensions gnome-shell-extension-dash-to-panel gnome-shell-extension-arc-menu
+    sudo apt-fast install -y gnome-tweaks gnome-shell-extensions
 fi
-sudo apt purge -y apport snaped
+sudo apt purge -y apport
 
-sudo add-apt-repository -y ppa:graphics-drivers/ppa
-echo '可安装的驱动：'
-ubuntu-drivers devices
-read -p '输入要安装的显卡驱动，按回车键跳过：' driver
-if [ -n "$driver" ]; then
-    sudo apt-fast install -y $driver
+read -p '要安装NVIDIA显卡驱动吗？(y/n)' install_nvidia
+if [[ "$install_nvidia" == "y" ]]; then
+    sudo add-apt-repository -y ppa:graphics-drivers/ppa
+    echo '可安装的驱动：'
+    ubuntu-drivers devices
+    read -p '输入要安装的显卡驱动，按回车键可跳过：' driver
+    if [ -n "$driver" ]; then
+        sudo apt-fast install -y $driver
+    fi
 fi
 
-echo '完成，请重启电脑！'
+if [[ "$XDG_CURRENT_DESKTOP" == "ubuntu:GNOME" ]]; then
+    echo 'GNOME插件下载地址：'
+    echo 'https://extensions.gnome.org/'
+    echo '推荐安装Dash to Panel和ArcMenu插件'
+
+fi
+
+echo '完成，要使显卡驱动生效，请重启电脑！'
